WIP: merge to Goerli testnet (#38)
* using geth

* remove ticker

* using generated token.json

* fix for goerli

* generate deploy info

* refactor: move event to delegate and fix erc20 (#42)

* fix base block

+++ contracts/IFluiDexDelegate.sol
@@ -24,7 +24,7 @@ interface IFluiDexDelegate {
     function depositERC20(
         IERC20 token,
         bytes32 to,
-        uint128 amount
+        uint256 amount
     ) external;
 
     /**
